refactor: reorganize type and plugin files
