+++ embeddable-say-editor/app/editor-element.ts
