+++ embeddable-say-editor/app/plugins/definitions/besluit-topic.gts
@@ -1,8 +1,8 @@
 import BesluitTopicDropdown from '@lblod/ember-rdfa-editor-lblod-plugins/components/besluit-topic-plugin/besluit-topic-toolbar-dropdown';
-import type { PluginInitializer } from '../../shared-types/embedded-plugin';
-import { mergeConfigs } from '../setup/defaults';
+import { mergeConfigs } from '../setup/defaults.ts';
 import type { TOC } from '@ember/component/template-only';
-import type { WidgetSignature } from '../../shared-types/widgets';
+import type { PluginInitializer } from '../embedded-plugin.ts';
+import type { WidgetSignature } from '../widgets.ts';
 
 const name = 'besluitTopic';
 export interface BesluitTopicConfig {
@@ -11,7 +11,7 @@ export interface BesluitTopicConfig {
   decisionUri?: string;
 }
 
-declare module '../../shared-types/plugin-registry' {
+declare module '../plugin-registry' {
   interface EmbeddedPlugins {
     [name]: typeof besluitTopic;
   }
