+++ embeddable-say-editor/app/plugins/definitions/editable-rdfa.ts
@@ -1,8 +1,8 @@
 import { editableNodePlugin } from '@lblod/ember-rdfa-editor/plugins/_private/editable-node';
-import type { PluginInitializer } from '../../shared-types/embedded-plugin';
+import type { PluginInitializer } from '../embedded-plugin.ts';
 
 const name = 'rdfaEditor';
-declare module '../../shared-types/plugin-registry' {
+declare module '../plugin-registry' {
   export interface EmbeddedPlugins {
     [name]: typeof setupEditableRdfaPlugin;
   }
