+++ embeddable-say-editor/app/plugins/definitions/html-edit.gts
@@ -1,8 +1,8 @@
-import type { PluginInitializer } from '../../shared-types/embedded-plugin';
+import type { PluginInitializer } from '../embedded-plugin';
 import HTMLEditorMenu from '@lblod/ember-rdfa-editor/components/plugins/html-editor/menu';
 const name = 'htmlEdit';
 
-declare module '../../shared-types/plugin-registry' {
+declare module '../plugin-registry' {
   export interface EmbeddedPlugins {
     [name]: typeof setupHtmlEdit;
   }
