+++ embeddable-say-editor/app/plugins/definitions/image.ts
@@ -3,7 +3,7 @@ import {
   imageView,
   imageWithConfig,
 } from '@lblod/ember-rdfa-editor/plugins/image';
-import type { PluginInitializer } from '../../shared-types/embedded-plugin';
+import type { PluginInitializer } from '../embedded-plugin.ts';
 import { mergeConfigs } from '../setup/defaults';
 export type ImagePluginConfig = {
   allowBase64Images: boolean;
@@ -12,7 +12,7 @@ export type ImagePluginConfig = {
 };
 
 const name = 'image';
-declare module '../../shared-types/plugin-registry' {
+declare module '../plugin-registry' {
   export interface PluginOptions {
     [name]?: Partial<ImagePluginConfig>;
   }
