+++ embeddable-say-editor/app/plugins/definitions/link.ts
@@ -3,11 +3,11 @@ import {
   link,
   linkPasteHandler,
 } from '@lblod/ember-rdfa-editor/plugins/link';
-import type { PluginInitializer } from '../../shared-types/embedded-plugin';
+import type { PluginInitializer } from '../embedded-plugin.ts';
 
 const name = 'link';
 
-declare module '../../shared-types/plugin-registry' {
+declare module '../plugin-registry' {
   export interface EmbeddedPlugins {
     [name]: typeof setupLinkPlugin;
   }
