+++ embeddable-say-editor/app/plugins/plugin-registry.ts
