+++ embeddable-say-editor/app/plugins/setup/defaults.ts
