+++ embeddable-say-editor/app/plugins/widgets.ts
@@ -1,6 +1,5 @@
 import type SayController from '@lblod/ember-rdfa-editor/core/say-controller';
 import type { ResolvedPNode } from '@lblod/ember-rdfa-editor/utils/_private/types';
-import type { EditorSetup } from '../app/setup/setup-plugins';
 import type { ComponentLike } from '@glint/template';
 import type { PluginName } from './embedded-plugin';
 import type {
@@ -8,6 +7,7 @@ import type {
   SidebarWidgets,
   ToolbarWidgets,
 } from './plugin-registry';
+import type { EditorSetup } from './setup/setup-plugins.ts';
 
 export type WidgetSignature<N extends PluginName | void = void> = {
   Args: {
