+++ embeddable-say-editor/app/render-editor-options.ts
