+++ embeddable-say-editor/app/utils/type-utils.ts
